Renamed directive file
