+++ core/templates/dev/head/pages/OppiaFooterDirective.ts
