+++ core/templates/dev/head/pages/donate/Donate.ts
@@ -19,7 +19,7 @@
 require('domain/utilities/UrlInterpolationService.ts');
 require('services/SiteAnalyticsService.ts');
 require('services/contextual/WindowDimensionsService.ts');
-require('pages/FooterDirective.ts');
+require('pages/OppiaFooterDirective.ts');
 
 oppia.controller('Donate', [
   '$http', '$scope', '$timeout', '$window', 'SiteAnalyticsService',
