+++ core/templates/dev/head/pages/thanks/Thanks.ts
@@ -17,7 +17,7 @@
  */
 
 require('components/background/BackgroundBannerDirective.ts');
-require('pages/FooterDirective.ts');
+require('pages/OppiaFooterDirective.ts');
 
 require('domain/utilities/UrlInterpolationService.ts');
